+++ libs/components/inline-form/projects/inline-form/src/modules/inline-form/fixtures/inline-form.fixture.module.ts
@@ -1,0 +1,26 @@
+import {
+  CommonModule
+} from '@angular/common';
+
+import {
+  NgModule
+} from '@angular/core';
+
+import {
+  SkyInlineFormModule
+} from '../inline-form.module';
+
+import {
+  SkyInlineFormFixtureComponent
+} from './inline-form.fixture';
+
+@NgModule({
+  declarations: [
+    SkyInlineFormFixtureComponent
+  ],
+  imports: [
+    CommonModule,
+    SkyInlineFormModule
+  ]
+})
+export class SkyInlineFormFixtureModule { }