+++ libs/components/inline-form/projects/inline-form/src/public-api.ts
@@ -1,0 +1,10 @@
+export * from './modules/inline-form/inline-form.module';
+export * from './modules/inline-form/types/inline-form-button-action';
+export * from './modules/inline-form/types/inline-form-button-config';
+export * from './modules/inline-form/types/inline-form-close-args';
+export * from './modules/inline-form/types/inline-form-config';
+export * from './modules/inline-form/types/inline-form-button-layout';
+
+// Components and directives must be exported to support Angular’s “partial” Ivy compiler.
+// Obscure names are used to indicate types are not part of the public API.
+export { SkyInlineFormComponent as λ1 } from './modules/inline-form/inline-form.component';