+++ libs/components/inline-form/src/app/public/public_api.ts
@@ -1,6 +1,0 @@
-export * from './modules/inline-form/inline-form.module';
-export * from './modules/inline-form/types/inline-form-button-action';
-export * from './modules/inline-form/types/inline-form-button-config';
-export * from './modules/inline-form/types/inline-form-close-args';
-export * from './modules/inline-form/types/inline-form-config';
-export * from './modules/inline-form/types/inline-form-button-layout';